+++ packages/axios-extension/src/base/service-provider.ts
@@ -62,16 +62,25 @@ export class ServiceProvider extends Axios implements ServiceProviderExtension {
     }
 
     /**
-     * Create a service instance for the given path and service class.
+     * Create a service instance for the given path, service class and public URL.
      *
      * @param path path of the service relative to the service provider
      * @param ServiceClass class type to be used to create an instance
      * @returns a service instance
      */
     protected createService<T extends Service>(path: string, ServiceClass: any): T {
-        const service = new ServiceClass(this.generateServiceConfig(path));
+        const service = new ServiceClass({ ...this.generateServiceConfig(path), publicUrl: this.publicUrl });
         service.log = this.log;
         service.interceptors = this.interceptors;
         return service;
+    }
+
+    /**
+     * Retrieve the public URL exposing the ABAP UI application.
+     *
+     * @returns string Axios Base URL
+     */
+    protected get publicUrl(): string {
+        return this.defaults.baseURL;
     }
 }